+++ webdev/lib/src/serve/webdev_server.dart
@@ -102,19 +102,19 @@ class WebDevServer {
         root: options.target,
       );
       dwds = await Dwds.start(
-        hostname: options.configuration.hostname,
-        assetReader: assetReader,
-        buildResults: filteredBuildResults,
-        chromeConnection: () async =>
-            (await Chrome.connectedInstance).chromeConnection,
-        logWriter: logWriter,
-        reloadConfiguration: options.configuration.reload,
-        serveDevTools:
-            options.configuration.debug || options.configuration.debugExtension,
-        verbose: options.configuration.verbose,
-        enableDebugExtension: options.configuration.debugExtension,
-        enableDebugging: options.configuration.debug,
-      );
+          hostname: options.configuration.hostname,
+          assetReader: assetReader,
+          buildResults: filteredBuildResults,
+          chromeConnection: () async =>
+              (await Chrome.connectedInstance).chromeConnection,
+          logWriter: logWriter,
+          reloadConfiguration: options.configuration.reload,
+          serveDevTools: options.configuration.debug ||
+              options.configuration.debugExtension,
+          verbose: options.configuration.verbose,
+          enableDebugExtension: options.configuration.debugExtension,
+          enableDebugging: options.configuration.debug,
+          expressionCompiler: null);
       pipeline = pipeline.addMiddleware(dwds.middleware);
       cascade = cascade.add(dwds.handler);
     }